refactor(insight): indep with visual-insight
