+++ packages/chart-advisor/src/insight/insightWorkers/vi-trend.ts
@@ -1,0 +1,5 @@
+import { Insight as VisualInsight } from 'visual-insights';
+import { visualInsightWorkerToAVAWorker } from './utils';
+import { Worker } from '.';
+
+export const trendIW: Worker = visualInsightWorkerToAVAWorker(VisualInsight.DefaultIWorker.trend);
